+++ spec/rubocop/cop/rails/unique_validation_without_index_spec.rb
@@ -89,6 +89,26 @@ RSpec.describe RuboCop::Cop::Rails::UniqueValidationWithoutIndex, :config do
       end
     end
 
+    context 'with a unique index and `check_constraint` that has `nil` first argument' do
+      let(:schema) { <<~RUBY }
+        ActiveRecord::Schema.define(version: 2020_02_02_075409) do
+          create_table "users", force: :cascade do |t|
+            t.string "account", null: false
+            t.index ["account"], name: "index_users_on_account", unique: true
+            t.check_constraint nil, 'expression', name: "constraint_name"
+          end
+        end
+      RUBY
+
+      it 'does not register an offense' do
+        expect_no_offenses(<<~RUBY)
+          class User
+            validates :account, uniqueness: true
+          end
+        RUBY
+      end
+    end
+
     context 'when the validation is for two columns' do
       context 'without proper index' do
         let(:schema) { <<~RUBY }